feat(aws-lambda) add support for forwarded request method/uri/body
New optional configuration properties:
- `forward_request_method`
- `forward_request_uri`
- `forward_request_headers`
- `forward_request_body`

If specified, the request body sent to the invoked Lambda will contain
the desired attributes of the client's request as a JSON payload.

If none of the above values are specified, the upstream body contains a
JSON payload which is merged from the request's body data and its
query string. This is to preserve backwards compatibility with previous
versions of this plugin.

Original patch from Andrei Kishkevich <v-akishkevich@expedia.com>
Reworked by Thibault Charbonnier <thibaultcha@me.com>

Signed-off-by: Thibault Charbonnier <thibaultcha@me.com>